+++ R/data.R
@@ -4,9 +4,11 @@
 #' Voxel coordinates for an organic shape
 #' 
 #' @examples
+#' head(organic_coords)
 #' cubes <- isocubesGrob(organic_coords, size = 2)
 #' grid::grid.newpage()
 #' grid::grid.draw(cubes)
+#' @family datasets
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 "organic_coords"
 
@@ -14,9 +16,23 @@
 #' Voxel coordinates for a sphere
 #' 
 #' @examples
+#' head(sphere_coords)
 #' cubes <- isocubesGrob(sphere_coords, size = 2)
 #' grid::grid.newpage()
 #' grid::grid.draw(cubes)
+#' @family datasets
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 "sphere_coords"
 
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Voxel coordinates for a sphere
+#' 
+#' @examples
+#' head(r_coords)
+#' cubes <- isocubesGrob(r_coords, size = 5, y = 0)
+#' grid::grid.newpage()
+#' grid::grid.draw(cubes)
+#' @family datasets
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+"r_coords"